👷Add rotating cards, more styles and card selection logic on checkout
