+++ src/Styles/CardImage.css
@@ -23,15 +23,6 @@
   overflow: hidden;
 }
 
-.card__gradient--1 {
-  background: #95e4c6;
-  background: linear-gradient(270deg, #f9c823 0%, #faa03c 33.65%, #fb7855 66.98%, #fc506e 100%);
-}
-.card__gradient--2 {
-  background: #780c67;
-  background: linear-gradient(68.44deg, #780c67 0%, #940967 41.98%, #b10667 73.75%, #d90266 100%);
-}
-
 .card__top-info {
   display: flex;
   justify-content: space-between;
@@ -84,9 +75,16 @@
 
 .card-funds__gradient {
   background: rgb(149, 228, 198);
-  background: linear-gradient(244.42deg, #363b3e 21.65%, #29efcb 70.95%);
+  /* background: linear-gradient(244.42deg, #363b3e 21.65%, #29efcb 70.95%); */
+  background: linear-gradient(
+    68.44deg,
+    #95f9c3 0%,
+    #73c9ac 18.54%,
+    #509995 41.98%,
+    #2e687d 66.98%,
+    #0b3866 100%
+  );
 }
-
 
 .card-funds__top-info {
   display: flex;
@@ -123,16 +121,34 @@
   background: radial-gradient(circle, rgba(0, 0, 0, 0.2) 0%, rgba(255, 255, 255, 0) 100%);
 }
 
-/* position */
-/* .cards__first {
-  grid-row: 3 / span 1;
-  z-index: 10;
+/* card gradients */
+.card__gradient--1 {
+  background: #95e4c6;
+  background: linear-gradient(270deg, #f9c823 0%, #faa03c 33.65%, #fb7855 66.98%, #fc506e 100%);
 }
-.cards__second {
-  grid-row: 2 / span 1;
-  z-index: 5;
+.card__gradient--2 {
+  background: #780c67;
+  background: linear-gradient(68.44deg, #780c67 0%, #940967 41.98%, #b10667 73.75%, #d90266 100%);
 }
-.cards__third {
-  grid-row: 1 / span 1;
-  z-index: 1;
-} */
+.card__gradient--3 {
+  background: #c05ede;
+  background: linear-gradient(
+    68.44deg,
+    #fffbaf 0%,
+    #eac7bf 28.44%,
+    #d593ce 53.96%,
+    #c05ede 80%,
+    rgba(171, 42, 237, 0.82) 100%
+  );
+}
+.card__gradient--4 {
+  background: #2e687d;
+  background: linear-gradient(
+    68.44deg,
+    #95f9c3 0%,
+    #73c9ac 18.54%,
+    #509995 41.98%,
+    #2e687d 66.98%,
+    #0b3866 100%
+  );
+}
